+++ tests/AbstractPdoEventStoreTest.php
@@ -19,7 +19,6 @@ use Prooph\EventStore\Exception\ConcurrencyException;
 use Prooph\EventStore\Metadata\FieldType;
 use Prooph\EventStore\Metadata\MetadataMatcher;
 use Prooph\EventStore\Metadata\Operator;
-use Prooph\EventStore\Pdo\Exception\JsonException;
 use Prooph\EventStore\Pdo\Exception\RuntimeException;
 use Prooph\EventStore\Stream;
 use Prooph\EventStore\StreamName;
@@ -482,34 +481,6 @@ abstract class AbstractPdoEventStoreTest extends AbstractEventStoreTest
         $streamEvents = $this->eventStore->load($streamName, 0, 10, $metadataMatcher);
 
         $this->assertCount(1, $streamEvents);
-    }
-
-    /**
-     * @test
-     * issue: https://github.com/prooph/pdo-event-store/issues/110
-     */
-    public function it_handles_invalid_json(): void
-    {
-        $this->expectException(JsonException::class);
-
-        $event = UserCreated::with(['name' => ['John', 'ßnow']], 1);
-        $event = $event->withAddedMetadata('key', 'value');
-
-        $streamName = new StreamName('Prooph\Model\User');
-        $stream = new Stream($streamName, new ArrayIterator([$event]), ['some' => ['metadata', 'as', 'well']]);
-
-        $this->eventStore->create($stream);
-
-        // Trigger an error when using an umlaut in the payload.
-        $this->connection->query('SET NAMES latin1');
-
-        $metadataMatcher = new MetadataMatcher();
-        $metadataMatcher = $metadataMatcher->withMetadataMatch(
-            'event_id', Operator::EQUALS(), $event->uuid()->toString(), FieldType::MESSAGE_PROPERTY()
-        );
-
-        $streamEvents = $this->eventStore->load($streamName, 1, null, $metadataMatcher);
-        $streamEvents->current(); // Trigger PdoStreamIterator.
     }
 
     /**
